Separate module with common testing logic
