--- loki-logback-appender/src/test/java/com/github/loki4j/SlowTests.java
@@ -1,3 +1,0 @@
-package com.github.loki4j;
-
-public interface SlowTests {}
