+++ loki-logback-appender/src/test/java/com/github/loki4j/logback/ApacheHttpAppenderTest.java
@@ -2,7 +2,8 @@ package com.github.loki4j.logback;
 
 import java.util.Random;
 
-import com.github.loki4j.logback.Generators.LokiHttpServerMock;
+import com.github.loki4j.testkit.dummy.LokiHttpServerMock;
+
 import static com.github.loki4j.logback.Generators.*;
 import static com.github.loki4j.logback.Loki4jAppenderTest.*;
