+++ loki-logback-appender/src/test/java/com/github/loki4j/logback/Generators.java
@@ -1,13 +1,9 @@
 package com.github.loki4j.logback;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.InetSocketAddress;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Iterator;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
@@ -18,7 +14,8 @@ import java.util.stream.StreamSupport;
 
 import com.github.loki4j.common.LogRecord;
 import com.github.loki4j.common.LokiResponse;
-import com.sun.net.httpserver.HttpServer;
+import com.github.loki4j.testkit.dummy.ExceptionGenerator;
+import com.github.loki4j.testkit.dummy.LokiHttpServerMock;
 
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.LoggerContext;
@@ -26,11 +23,9 @@ import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.classic.spi.LoggingEvent;
 import ch.qos.logback.classic.spi.ThrowableProxy;
 
-public class Generators {
+import static com.github.loki4j.testkit.dummy.Generators.genMessage;
 
-    public static Exception exception(String message) {
-        return new RuntimeException(message);
-    }
+public class Generators {
 
     public static String batchToString(LogRecord[] batch) {
         var s = new StringBuilder();
@@ -251,30 +246,11 @@ public class Generators {
                 Level.ERROR,
                 "test.TestApp",
                 "thread-" + (i % 8),
-                String.format("#%s - %s", i, "Error occured"),
-                exception(msg)));
+                String.format("#%s - %s", i, "Error occurred"),
+                ExceptionGenerator.exception(msg)));
         }
 
         return events.toArray(new LoggingEvent[0]);
-    }
-
-    private static String genMessage(int maxWords) {
-        ThreadLocalRandom rnd = ThreadLocalRandom.current();
-
-        StringBuilder msg = new StringBuilder();
-        int words = rnd.nextInt(1, maxWords);
-        for (int i = 0; i < words; i++) {
-            int letters = rnd.nextInt(1, 20);
-            for (int j = 0; j < letters; j++) {
-                msg.append(rnd.nextFloat() < 0.1
-                    ? (char)('A' + rnd.nextInt('Z' - 'A'))
-                    : (char)('a' + rnd.nextInt('z' - 'a')));
-            }
-            msg.append(rnd.nextFloat() < 0.05
-                ? '\n'
-                : ' ');
-        }
-        return msg.toString();
     }
 
     public static LoggingEvent loggingEvent(
@@ -333,44 +309,4 @@ public class Generators {
         }
     }
 
-
-    public static class LokiHttpServerMock {
-        public List<byte[]> batches = new ArrayList<>();
-        public volatile byte[] lastBatch;
-
-        private final HttpServer server;
-
-        public LokiHttpServerMock(int port) throws IOException {
-            server = HttpServer.create(new InetSocketAddress(port), 0);
-            server.createContext("/loki/api/v1/push", httpExchange -> {
-                try (var is = httpExchange.getRequestBody()) {
-                    lastBatch = getBytesFromInputStream(is); //is.readAllBytes();
-                    batches.add(lastBatch);
-                }
-                httpExchange.sendResponseHeaders(204, -1);
-            });
-        }
-
-        public static byte[] getBytesFromInputStream(InputStream is) throws IOException {
-            ByteArrayOutputStream os = new ByteArrayOutputStream();
-            byte[] buffer = new byte[0xFFFF];
-            for (int len = is.read(buffer); len != -1; len = is.read(buffer)) {
-                os.write(buffer, 0, len);
-            }
-            return os.toByteArray();
-        }
-
-        public void start() {
-            new Thread(server::start).start();
-        }
-
-        public void stop() {
-            server.stop(0);
-        }
-
-        public void reset() {
-            batches.clear();
-            lastBatch = null;
-        }
-    }
 }
