+++ loki-logback-appender/src/test/java/com/github/loki4j/logback/integration/GrafanaCloudTest.java
@@ -4,6 +4,7 @@ import static com.github.loki4j.logback.Generators.*;
 import static org.junit.Assert.*;
 
 import com.github.loki4j.logback.AbstractHttpSender;
+import com.github.loki4j.testkit.categories.IntegrationTests;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
