--- loki-logback-appender/src/test/java/com/github/loki4j/logback/integration/IntegrationTests.java
@@ -1,5 +1,0 @@
-package com.github.loki4j.logback.integration;
-
-import com.github.loki4j.SlowTests;
-
-public interface IntegrationTests extends SlowTests {};
