+++ loki-logback-appender/src/test/java/com/github/loki4j/logback/integration/LargeBatchSendTest.java
@@ -3,6 +3,8 @@ package com.github.loki4j.logback.integration;
 import static com.github.loki4j.logback.Generators.*;
 import static org.junit.Assert.*;
 
+import com.github.loki4j.testkit.categories.IntegrationTests;
+
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
