+++ loki-logback-appender/src/test/java/com/github/loki4j/logback/performance/EncodersTest.java
@@ -8,7 +8,9 @@ import java.util.stream.Stream;
 
 import com.github.loki4j.common.LogRecord;
 import com.github.loki4j.logback.AbstractLoki4jEncoder;
-import com.github.loki4j.logback.performance.Benchmarker.Benchmark;
+import com.github.loki4j.testkit.benchmark.Benchmarker;
+import com.github.loki4j.testkit.benchmark.Benchmarker.Benchmark;
+import com.github.loki4j.testkit.categories.PerformanceTests;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
