--- loki-logback-appender/src/test/java/com/github/loki4j/logback/performance/PerformanceTests.java
@@ -1,5 +1,0 @@
-package com.github.loki4j.logback.performance;
-
-import com.github.loki4j.SlowTests;
-
-public interface PerformanceTests extends SlowTests {};
