+++ settings.gradle
@@ -4,6 +4,7 @@ import static groovy.io.FileType.*
 rootProject.name = 'loki4j'
 
 include 'common'
+include 'testkit'
 include 'loki-logback-appender'
 
 def tempProjects = new File(rootDir, 'temp')
