+++ testkit/src/main/java/com/github/loki4j/testkit/categories/SlowTests.java
@@ -1,0 +1,3 @@
+package com.github.loki4j.testkit.categories;
+
+public interface SlowTests {}
