+++ testkit/src/main/java/com/github/loki4j/testkit/dummy/ExceptionGenerator.java
@@ -1,0 +1,9 @@
+package com.github.loki4j.testkit.dummy;
+
+public class ExceptionGenerator {
+
+    public static Exception exception(String message) {
+        return new RuntimeException(message);
+    }
+    
+}
